Print MongoDB collection BeforeScenario
