+++ tests/IntegrationTests/TaskManager.IntegrationTests/StepDefinitions/CommonApiDefinitions.cs
@@ -18,6 +18,7 @@ using System.Net;
 using FluentAssertions;
 using Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.POCO;
 using Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.Support;
+using MongoDB.Driver;
 using Newtonsoft.Json;
 
 namespace Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.StepDefinitions
@@ -25,9 +26,12 @@ namespace Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.StepDefiniti
     [Binding]
     public class CommonApiDefinitions
     {
+        private MongoClientUtil MongoClient { get; }
+
         public CommonApiDefinitions(ObjectContainer objectContainer)
         {
             ApiHelper = objectContainer.Resolve<ApiHelper>();
+            MongoClient = objectContainer.Resolve<MongoClientUtil>();
         }
 
         [Given(@"I have a TaskManager endpoint (.*)")]
@@ -46,6 +50,9 @@ namespace Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.StepDefiniti
         [Then(@"I will get a (.*) response")]
         public void ThenIWillGetAResponse(string expectedCode)
         {
+            var names = MongoClient.Database.ListCollectionNames().ToList();
+            Console.WriteLine("Mongo Collections={0}", names.Count);
+            names.ForEach(p => Console.WriteLine("Mong Collection={0}", p));
             ApiHelper.Response?.StatusCode.Should().Be((HttpStatusCode)Enum.Parse(typeof(HttpStatusCode), expectedCode));
         }
