+++ tests/IntegrationTests/TaskManager.IntegrationTests/Support/MongoClientUtil.cs
@@ -25,7 +25,7 @@ namespace Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.Support
     public class MongoClientUtil
     {
         private MongoClient Client { get; set; }
-        private IMongoDatabase Database { get; set; }
+        internal IMongoDatabase Database { get; set; }
         private IMongoCollection<TaskDispatchEventInfo> TaskDispatchEventInfoCollection { get; set; }
         private RetryPolicy RetryMongo { get; set; }
         private RetryPolicy<List<TaskDispatchEventInfo>> RetryTaskDispatchEventInfo { get; set; }
@@ -70,6 +70,13 @@ namespace Monai.Deploy.WorkflowManager.TaskManager.IntegrationTests.Support
         public void DropDatabase(string dbName)
         {
             Client.DropDatabase(dbName);
+        }
+
+        internal void ListAllCollections(ISpecFlowOutputHelper outputHelper, string testFeature)
+        {
+            var collections = Database.ListCollectionNames().ToList();
+            outputHelper.WriteLine($"MongoDB collections found in test feature '{testFeature}': {collections.Count}");
+            collections.ForEach(p => outputHelper.WriteLine($"- Collection: {p}"));
         }
     }
 }
