+++ tests/IntegrationTests/WorkflowExecutor.IntegrationTests/StepDefinitions/PayloadApiStepDefinitions.cs
@@ -100,10 +100,6 @@ namespace Monai.Deploy.WorkflowManager.WorkflowExecutor.IntegrationTests.StepDef
         [Then(@"I can see no Payloads are returned")]
         public void ThenICanSeeNoPayloadsAreReturned()
         {
-            var names = MongoClient.Database.ListCollectionNames().ToList();
-            Console.WriteLine("Mongo Collections={0}", names.Count);
-            names.ForEach(p => Console.WriteLine("Mong Collection={0}", p));
-            
             var result = ApiHelper.Response.Content.ReadAsStringAsync().Result;
             var payloads = JsonConvert.DeserializeObject<PagedResponse<List<Payload>>>(result);
