+++ tests/IntegrationTests/WorkflowExecutor.IntegrationTests/StepDefinitions/WorkflowRequestStepDefinitions.cs
@@ -162,10 +162,6 @@ namespace Monai.Deploy.WorkflowManager.IntegrationTests.StepDefinitions
         [Then(@"No workflow instances will be created")]
         public void ThenTheWorkflowWillNotTriggerAnyNewWorkflowInstances()
         {
-            var names = MongoClient.Database.ListCollectionNames().ToList();
-            Console.WriteLine("Mongo Collections={0}", names.Count);
-            names.ForEach(p => Console.WriteLine("Mong Collection={0}", p));
-            
             foreach (var workflowRevision in DataHelper.WorkflowRevisions)
             {
                 for (int i = 0; i < 5; i++)
